RoboSimSimulator and 2DRobotSimulator ported to new API; 2DRobotSimulator exit problems fixed
git-svn-id: https://svn.code.sf.net/p/plexil/code/branches/plexil-3-new-exp@3850 d2be67b0-814c-0410-949f-d215b6b29573

+++ examples/robosim/2DRobotSimulator/Robot.hh
@@ -1,4 +1,4 @@
-/* Copyright (c) 2006-2008, Universities Space Research Association (USRA).
+/* Copyright (c) 2006-2014, Universities Space Research Association (USRA).
 *  All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without
@@ -29,8 +29,6 @@
 
 #include "RobotBase.hh"
 
-#include <vector>
-
 #include "ThreadMutex.hh"
 
 class IpcRobotAdapter;
@@ -60,7 +58,7 @@ public:
   double determineEnergySourceLevel();
   double determineGoalLevel();
 
-  const std::vector<double> processCommand(const std::string& cmd, double parameter = 0.0);
+  const std::vector<double> processCommand(const std::string& cmd, int32_t parameter = 0);
 
 private:
